requirements.txt
Django==5.2.6
Pillow==10.4.0
django-crispy-forms==2.3
crispy-bootstrap5==2024.2
django-extensions==3.2.3
djangorestframework==3.15.2
django-widget-tweaks==1.5.0
pyotp==2.9.0
qrcode==7.4.2

# Production dependencies
# psycopg2-binary==2.9.9  # Commented out: fails on Windows Python 3.13 - use for production with PostgreSQL
redis==5.0.8
gunicorn==21.2.0
whitenoise==6.6.0

# Optional production enhancements
django-storages==1.14.4
boto3==1.34.162
python-decouple==3.8

# Development dependencies (optional)
django-debug-toolbar==4.4.6
coverage==7.6.1
pytest==8.3.2
pytest-django==4.8.0
factory-boy==3.3.0

# Security and monitoring
django-csp==3.8  # Content Security Policy headers (P1)
sentry-sdk==2.13.0
django-cors-headers==4.4.0

# Payment processing
stripe==11.1.1

# Push notifications
pywebpush==2.0.1

# HTTP library (if not already included by other deps)
requests==2.32.3